Add files via upload
Preparing a monthly sales report
Overview
In the exercise Preparing a monthly Sales Report, you were asked to put into practice what you have learned about the five core functions SUM, AVERAGE, COUNT, MAX, and MIN. 
Your task in this exercise was to create appropriate formulas to identify the following information: 
•	The total revenue from the A2 Mountain Bike Frame sales for April.
•	The number of frames sold.
•	The lowest and highest daily sales figures.
•	The number of days in the month.
•	And the overall daily sales average.
This reading provides you with a step-by-step guide for identifying these results. It also includes screenshots that you can compare against your own work. 
You can review the formula creation techniques for these functions in the videos What is a function formula and how to write it and Using the insert function and in the reading The AutoSum shortcut.
Step 1: Download the file.
•	You began by downloading and opening the Microsoft Excel workbook Monthly sales report.xlsx. The workbook contained one worksheet called A2Mountain Bike Frames, which included information on the daily sales of this product during the month of April. You were asked to create formulas in the cell range C35 to C40. You confirmed that the Number format applied to these cells was the General format by checking the Number section of the Home ribbon.
 
Step 2: Creating the Calculations.
1.	The first formula that you were asked to create in cell C35, calculates the total revenue for April. With the cursor positioned on cell C35, you created this formula using the SUM function to calculate the total revenue for the figures in the cell range E4 to E33. The formula to generate this result should read:
=SUM(E4:E33)
You could have created this formula by typing it directly into C35. You could also have used the Insert function wizard or the AutoSum shortcut list. With both shortcut methods, Excel initially suggests the cell range C4:C33 because these values are directly above the cursor position. If you used one of these methods to create the formula, you would have had to manually select the revenue values in column E.
The formula generated a result in cell C35 of $23,059,600.
The format on C35 is automatically changed to the Accounting format because all the values in the cell range E4 to E33 were formatted in that way.
 
1.	You then had to work out how many A2Mountain Bike Frames were sold in April. You created this formula in C36 using the SUM function to calculate the total number of units in cells C4 to C33. 
The formula to generate this result reads:
=SUM(C4:C33)
If you used the AutoSum shortcut to create this calculation, Excel may have suggested only one cell reference in column C. This happened because there is a formula in cell C35 and cell C34 is blank. If you used this method, then you would have had to manually select the correct cell range C4:C33 for the calculation.
The result in C36 was 115,298.
 
1.	The next piece of information that you needed for the report was the day when the lowest number of units were sold. You were asked to note the date and the number of units sold and type the date in cell D37.
To identify the lowest result in the Number of units sold column, you created a formula in C37 using the MIN function. The cell range for the MIN function argument is C4 to C33. The formula to generate this result is:
=MIN(C4:C33)
The lowest number of units sold was 2,560 on April 30, 2023.
 
1.	You also had to create a formula in cell C38 to identify the day the most units were sold. You were asked to make a note of the date and the amount and type the date in cell D38.
To identify the largest result in the Number of units sold column, you created a formula in C38 using the MAX function. The cell range for the MAX function argument is C4 to C33. The formula to generate this result was as follows:
=MAX(C4:C33)
The largest number of units sold was 4,921 on April 16, 2023.
 
1.	The report also had to show how many days there were in the month and this result had to be calculated in cell C39.
Dates are always typed into a spreadsheet as numbers separated by the forward slash character. This character is used to separate the month, day, and year. Excel stores dates as numeric content.
The dates in the cell range B4 to B33 had been entered in the MM/DD/YY format but had also been customized to appear as a mixture of text and numbers. However, you were able confirm that these are still numeric entries by selecting any cell in the range and then checking the formula bar. 
Because of the way that the dates were displayed, you might have used the COUNTA function, which counts any content, including text and numeric. In this case, the COUNT function would also have given a correct result since the dates are stored as numeric data. 
The formula to count the number of days could have been as follows:
=COUNT(B4:B33)
 
Or you could also have use the following formula:
=COUNTA(B4:B33)
 
With both versions of the formula, the number of days in the month was calculated as 30.
1.	You also needed to calculate the average daily totals for these product sales by creating a formula in cell C40 .
Again, if you used the Insert function or the AutoSum shortcut to create this formula, Excel would have suggested an incorrect cell range. You would then have to manually select the correct range of E4 to E33.
The formula to generate the required result was:
=AVERAGE(E4:E33)
The average daily sales dollar amount for this product was $768,653. Excel automatically formatted the result with the Accounting format and displayed it as dollars.
 
Conclusion